Fixed API backend problems.
This isn't really what this branch was intended for, but that doesn't
matter that much. The basic point is that I need to do some more work
with the backend/state/ngrx/mechanics before I really feel comfortable
working on the UI. I want to be able to load current configuration,
state, etc, and that just isn't possible (or easy to implement) right
now. I think my next step is to work on completely refactoring the
firebase functions system to use express and be better split between
different files.